+++ CHANGES.md
@@ -8,6 +8,13 @@ Version 0.1.1
 
 To be released.
 
+ -  `BencodexDictionary` and `RecordView` now shows easier inspection
+     representions on the eyes both on Deno ([`Deno.inspect()`]) and
+     Node.js ([`util.inspect()`]).
+
+[`Deno.inspect()`]: https://deno.land/api?s=Deno.inspect
+[`util.inspect()`]: https://nodejs.org/api/util.html#utilinspectobject-options
+
 
 Version 0.1.0
 -------------